feat(TD-7): Refactor code according to Next.js conventions
