+++ modules/to-do/components/AddTask/AddTask.types.ts
@@ -1,0 +1,4 @@
+export interface IAddTaskProps {
+    onSubmit: (task: { title: string }) => void;
+  }
+
