+++ modules/to-do/components/TaskList/TaskList.tsx
@@ -1,0 +1,16 @@
+import { ITaskListProps } from "./TaskList.types";
+
+export const TaskList = ({ tasks }: ITaskListProps) => {
+  return (
+    <ul className="mt-4 space-y-2 flex flex-col items-center">
+      {tasks.map((task) => (
+        <li
+          key={task.id}
+          className="p-4 border rounded bg-gray-200 shadow-sm w-64 text-black"
+        >
+          <h3 className="font-bold">{task.title}</h3>
+        </li>
+      ))}
+    </ul>
+  );
+};
