+++ modules/to-do/components/TaskList/TaskList.types.ts
@@ -1,0 +1,5 @@
+import { Task } from "@/shared/typedefs";
+
+export interface ITaskListProps {
+  tasks: Task[];
+}
