+++ shared/components/ui/index.ts
@@ -1,0 +1,3 @@
+export * from "./button";
+export * from "./input";
+export * from "./textarea";
