+++ shared/typedefs/enums.ts
@@ -1,0 +1,1 @@
+export {};
