+++ shared/typedefs/interfaces.ts
@@ -3,3 +3,6 @@ export interface Task {
     title: string;
     completed: boolean;
   }
+  
+
+
